+++ cloudplatform/connectivity-apache-httpclient5/src/test/java/com/sap/cloud/sdk/cloudplatform/connectivity/ApacheHttpClient5HeaderTest.java
@@ -58,13 +58,21 @@ class ApacheHttpClient5HeaderTest
     @RequiredArgsConstructor
     enum TestAssertion
     {
-        NONE(List.of("Connection", "Host", "User-Agent", "Accept-Encoding"), Map.of()),
-        KEEP_ALIVE(List.of("Connection", "Host", "User-Agent", "Accept-Encoding"), Map.of("Connection", "keep-alive")),
+        NONE(
+            List.of("Connection", "Host", "User-Agent", "Accept-Encoding"),
+            List.of("Connection", "Upgrade"),
+            Map.of()),
+        KEEP_ALIVE(
+            List.of("Connection", "Host", "User-Agent", "Accept-Encoding"),
+            List.of("Upgrade"),
+            Map.of("Connection", "keep-alive")),
         UPGRADE(
             List.of("Connection", "Host", "User-Agent", "Accept-Encoding", "Upgrade"),
+            List.of(),
             Map.of("Connection", "Upgrade", "Upgrade", "TLS/1.2"));
 
         private final List<String> headersAllowed;
+        private final List<String> headersForbidden;
         private final Map<String, String> headersRequired;
     }
 
@@ -86,7 +94,7 @@ class ApacheHttpClient5HeaderTest
             new TestCase(TestDestination.TLS_VERSION, ENABLED, TestAssertion.KEEP_ALIVE),
 
             // Disabled HTTP/TLS upgrade
-            new TestCase(TestDestination.INTERNET, DISABLED, TestAssertion.NONE),
+            new TestCase(TestDestination.INTERNET, DISABLED, TestAssertion.KEEP_ALIVE),
             new TestCase(TestDestination.PROXY, DISABLED, TestAssertion.NONE),
             new TestCase(TestDestination.ON_PREMISE, DISABLED, TestAssertion.NONE),
             new TestCase(TestDestination.TLS_VERSION, DISABLED, TestAssertion.KEEP_ALIVE),
@@ -104,11 +112,12 @@ class ApacheHttpClient5HeaderTest
     {
         stubFor(get(anyUrl()).willReturn(ok()));
 
-        var sut = new ApacheHttpClient5FactoryBuilder().build();
+        var sut = new ApacheHttpClient5FactoryBuilder().tlsUpgrade(testCase.tlsFlag).build();
         var dest = testCase.destination.destinationBuilder.apply(server.getHttpBaseUrl());
         sut.createHttpClient(dest).execute(new HttpGet("/foo"), new BasicHttpClientResponseHandler());
 
         var request = getRequestedFor(anyUrl()).andMatching(allowedHeaders(testCase.assertion.headersAllowed));
+        testCase.assertion.headersForbidden.forEach(request::withoutHeader);
         testCase.assertion.headersRequired.forEach(( k, v ) -> request.withHeader(k, equalTo(v)));
         verify(request);
     }