+++ release_notes.md
@@ -8,7 +8,11 @@
 
 ### 🔧 Compatibility Notes
 
-- 
+- The TLS `Upgrade` header changes thanks to [Apache httpclient5 5.4.2](https://github.com/apache/httpcomponents-client/commit/5ab09ea39fed1c39ea35905532ba1567c785330a)
+  - `TlsUpgrade.DISABLED` no changes
+  - `TlsUpgrade.ENABLED` will not send the `Upgrade` header for non-proxy connections anymore
+  - `TlsUpgrade.AUTOMATIC` **Default behaviour** will not send the `Upgrade` header anymore
+    - Except for `proxyType(ProxyType.INTERNET)`
 
 ### ✨ New Functionality
 
@@ -20,5 +24,4 @@
 
 ### 🐛 Fixed Issues
 
-- Fix OData `VDMObject.getChangedFields()` to use `BigDecimal.compareTo()` instead of `BigDecimal.equals()`.
-  - Example: a `BigDecimal` field updated from `1` to `1.0` will not be considered as changed anymore.
+-